rename appstack -> app_stack module
